+++ cphasing/_config.py
@@ -9,3 +9,6 @@ READ_COORD_DTYPE = "uint32"
 READ_LENGTH_DTYPE = "uint32"
 MAPPING_QUALITY_DTYPE = "uint32"
 PERCENTAGE_DTYPE = "float32"
+
+## Datatype of HyperGraph order
+HYPERGRAPH_ORDER_DTYPE = "int8"